Move deck production out of Main
