+++ src/main/java/hanziToAnki/ArgParser.java
@@ -4,8 +4,26 @@ import org.apache.commons.io.FilenameUtils;
 
 import java.util.ArrayList;
 import java.util.List;
+import java.util.stream.Collectors;
+import java.util.stream.Stream;
 
 public class ArgParser {
+
+    public static void printUsage() {
+        System.out.println("Usage: java hanziToAnki.Main [OPTIONS] filename");
+        System.out.println("options:");
+        System.out.println(
+                "\t-w --word-list:\tRead from an input file containing a list of words, separated"
+                        + " by line breaks. Without this flag, individual characters are extracted.");
+        System.out.println("\t-s --single-characters:\tdictHandler.Extract only single characters from the file.");
+        System.out.println("\t-hsk <hsk level> Remove any words in any HSK levels up to and including"
+                + " the given one.");
+        System.out.println("\t-o <output filename> Override the default output file name");
+        System.out.println("\t-f --format <output format> Override the default output file name\n" +
+                "\t\tChoices are: " + Stream.of(OutputFormat.values())
+                .map(OutputFormat::name)
+                .collect(Collectors.joining(", ")));
+    }
 
     public record ParsedArgs(ExportOptions options, List<String>fileNames, String outputFileName) {}
